requirements.txt
boto3
Flask
PyGithub
requests